Add files via upload
This is a new data file from preliminary experiments conducted in 2018-2019. There is only one medium concentration of 100 % and three adaptation temperatures. Not everything was measured (in particular we did not measure cell shape and volume from the videos), and we were still setting up the protocol, meaning that for instance the microscope slide was not already at temperature before pipetting the cells. The recording under the microscope lasted for 20 minutes in each condition, meaning that we can look at changes of speed over time.